backend/app/core/services/pricing_service.py
import httpx
import asyncio
from typing import Dict, Optional
from datetime import datetime, timedelta
import re
from bs4 import BeautifulSoup

class PricingService:
    def __init__(self):
        self.cache: Dict[str, Dict] = {}
        self.cache_duration = timedelta(hours=24)  # Cache pricing for 24 hours
        self.last_update: Dict[str, datetime] = {}
        
        # Provider pricing URLs
        self.pricing_urls = {
            "openai": "https://openai.com/api/pricing",
            "anthropic": "https://www.anthropic.com/pricing",
            "xai": "https://xai.com/pricing"  # Replace with actual XAI pricing URL
        }
    
    def get_pricing(self, provider: str) -> Dict[str, float]:
        """Get pricing information for a specific provider."""
        current_time = datetime.now()
        
        # Check if we have valid cached pricing
        if (provider in self.cache and 
            provider in self.last_update and 
            current_time - self.last_update[provider] < self.cache_duration):
            return self.cache[provider]
        
        # Fetch fresh pricing
        pricing = self._fetch_pricing(provider)
        self.cache[provider] = pricing
        self.last_update[provider] = current_time
        return pricing
    
    def _fetch_pricing(self, provider: str) -> Dict[str, float]:
        """Fetch pricing information from provider's website."""
        if provider not in self.pricing_urls:
            raise ValueError(f"Unknown provider: {provider}")
        return self._get_default_pricing(provider)
    
    async def _parse_openai_pricing(self, html: str) -> Dict[str, float]:
        """Parse OpenAI pricing from their website."""
        # TODO: Implement OpenAI-specific parsing logic once their pricing page is available
        return self._get_default_pricing("openai")
    
    async def _parse_anthropic_pricing(self, html: str) -> Dict[str, float]:
        """Parse Anthropic pricing from their website."""
        # TODO: Implement Anthropic-specific parsing logic once their pricing page is available
        return self._get_default_pricing("anthropic")
    
    async def _parse_xai_pricing(self, html: str) -> Dict[str, float]:
        """Parse XAI pricing from their website."""
        # TODO: Implement XAI-specific parsing logic once their pricing page is available
        return self._get_default_pricing("xai")
    
    def _get_default_pricing(self, provider: str) -> Dict[str, float]:
        """Return default pricing if website parsing fails."""
        # TODO: Write a daily cron to update the pricing
        defaults = {
            "openai": {
                "prompt_price_per_1k": 0.03,
                "completion_price_per_1k": 0.06
            },
            "anthropic": {
                "prompt_price_per_1k": 0.015,
                "completion_price_per_1k": 0.075
            },
            "xai": {
                "prompt_price_per_1k": 0.02,
                "completion_price_per_1k": 0.04
            }
        }
        return defaults.get(provider, {"prompt_price_per_1k": 0.0, "completion_price_per_1k": 0.0}) 