backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai==1.78.0
anthropic>=0.5.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
httpx>=0.24.0
psycopg2-binary>=2.9.9  # Required for SQLAlchemy with PostgreSQL
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9 
pydantic[email]
bcrypt